refactor(neon_files): externalise files widgets
